+++ marshmallow/utils.py
@@ -13,7 +13,6 @@ from pprint import pprint as py_pprint
 
 from marshmallow.compat import OrderedDict
 
-
 def is_iterable_but_not_string(obj):
     '''Return True if ``obj`` is an iterable object that isn't a string.'''
     return hasattr(obj, "__iter__") and not hasattr(obj, "strip")
@@ -75,7 +74,6 @@ def to_marshallable_type(obj, field_names=None):
         attrs = set(dir(obj))
     return dict([(attr, getattr(obj, attr, None)) for attr in attrs
                 if not attr.startswith("__") and not attr.endswith("__")])
-
 
 
 def pprint(obj, *args, **kwargs):